requirements.txt
astroid==1.6.5
backports.functools-lru-cache==1.5
beautifulsoup4==4.7.0
blinker==1.4
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
Click==7.0
configparser==3.5.0
Django==1.11.16
dnspython==1.16.0
enum34==1.1.6
Faker==1.0.1
Flask==1.0.2
Flask-Cors==3.0.7
Flask-Mail==0.9.1
Flask-SocketIO==3.3.2
geographiclib==1.49
geopy==1.18.1
gevent==1.4.0
google-images-download==2.5.0
greenlet==0.4.15
idna==2.8
ipaddress==1.0.22
isort==4.3.4
itsdangerous==1.1.0
javascript==2015.32.59
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
node==0.9.23
npm==0.1.1
odict==1.6.2
optional-django==0.3.0
Pillow==5.4.1
plumber==1.5
pm==0.1.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.7
pylint==1.9.3
python-dateutil==2.7.5
python-engineio==3.1.2
python-socketio==3.0.0
pytz==2018.7
react==4.3.0
requests==2.21.0
schedule==0.5.0
selenium==3.141.0
singledispatch==3.4.0.3
six==1.12.0
soupsieve==1.6.1
termcolor==1.1.0
text-unidecode==1.2
urllib3==1.24.1
virtualenv==16.1.0
Werkzeug==0.14.1
wrapt==1.10.11
zope.component==4.5
zope.deferredimport==4.3
zope.deprecation==4.3.0
zope.event==4.4
zope.hookable==4.2.0
zope.interface==4.6.0
zope.lifecycleevent==4.3
zope.proxy==4.3.1
